Extract 'get' to its own file
